requirements.txt
speech_recognition
win32com.client
random
webbrowser
datetime
jikanpy